requirements.txt
streamlit==1.33.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
google-genai==0.3.0
pydantic==2.6.4
plotly==5.20.0
python-dotenv==1.0.1